organised platform_interface module
