+++ packages/cloud_firestore/cloud_firestore_platform_interface/lib/src/method_channel_collection_reference.dart
@@ -7,7 +7,7 @@ part of cloud_firestore_platform_interface;
 /// A CollectionReference object can be used for adding documents, getting
 /// document references, and querying for documents (using the methods
 /// inherited from [Query]).
-class MethodChannelCollectionReference extends QueryPlatform  implements CollectionReference {
+class MethodChannelCollectionReference extends MethodChannelQuery  implements CollectionReference {
   MethodChannelCollectionReference(FirestorePlatform firestore, List<String> pathComponents)
       : super(firestore: firestore,pathComponents: pathComponents);
