+++ packages/cloud_firestore/cloud_firestore_platform_interface/lib/src/platform_interface/query_snapshot.dart
@@ -5,7 +5,7 @@
 part of cloud_firestore_platform_interface;
 
 /// A QuerySnapshot contains zero or more DocumentSnapshot objects.
-class QuerySnapshot {
+abstract class QuerySnapshot {
   
   QuerySnapshot(this.documents, this.documentChanges, this.metadata);
   /// Gets a list of all the documents included in this snapshot
